feat(settings): restructurer en 4 onglets (Général, Instructions, Connaissances, Avancé)
